@@ -58,7 +58,7 @@ class _AdminDashboardPageState extends State<AdminDashboardPage> {
               unselectedItemColor:
                   Theme.of(
                     context,
-                  ).textTheme.bodyMedium?.color?.withValues(alpha: 0.6) ??
+                  ).textTheme.bodyMedium?.color?.withValues(alpha:  0.6) ??
                   AppColors.textSecondary,
               items: NavigationMenuConfig.items
                   .map(
